# openapi.yaml
openapi: 3.0.3
info:
  title: ExamFlow Shared API Contract
  version: 0.2.0
  description: >
    Shared React-facing contract for the Ruby primary API and C# backup API.
    Both providers must expose these paths under /api/v1 and use the same SQL Server data.
servers:
  - url: http://localhost:3000/api/v1
    description: Ruby primary API
  - url: http://localhost:5014/api/v1
    description: C# backup API
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate a user against NguoiDung.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required: [password]
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials.
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create a SQL-backed user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Signup succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "409":
          description: Email or generated student code already exists.
  /nguoi_dung:
    get:
      tags: [Accounts]
      summary: List user accounts.
      responses:
        "200":
          description: User list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NguoiDung"
    post:
      tags: [Accounts]
      summary: Create a user account.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nguoi_dung:
                  $ref: "#/components/schemas/AccountWrite"
      responses:
        "201":
          description: Created user account.
  /nguoi_dung/{id}:
    get:
      tags: [Accounts]
      summary: Get a user account.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: User account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NguoiDung"
        "404":
          description: User not found.
    patch:
      tags: [Accounts]
      summary: Update a user account.
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nguoi_dung:
                  $ref: "#/components/schemas/AccountWrite"
      responses:
        "200":
          description: Updated user account.
    delete:
      tags: [Accounts]
      summary: Delete a user account.
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Deleted user account.
        "409":
          description: User is referenced by operational data.
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Required before unsafe write retry is enabled across Ruby/C#.
  schemas:
...
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/NguoiDung"
        authSource:
          type: string
        generatedCode:
          type: string
    SignupRequest:
      type: object
      properties:
        HoTen:
          type: string
        Email:
          type: string
        MatKhau:
          type: string
        VaiTroID:
          type: integer
        MaDinhDanh:
          type: string
        MaSinhVien:
          type: string
      required: [HoTen, Email, MatKhau]
    NguoiDung:
      type: object
      properties:
        NguoiDungID:
          type: integer
        Email:
          type: string
        HoTen:
          type: string
        VaiTroID:
          type: integer
        TenVaiTro:
          type: string
        TrangThai:
          oneOf:
            - type: integer
            - type: boolean
        MaSinhVien:
          type: string
        TaoLuc:
          type: string
          format: date-time
        CapNhatLuc:
          type: string
          format: date-time
    AccountWrite:
      type: object
      properties:
        Email:
          type: string
        HoTen:
          type: string
        MatKhau:
          type: string
        MatKhauHash:
          type: string
        VaiTroID:
          type: integer
        TrangThai:
          oneOf:
            - type: integer
            - type: boolean
        MaSinhVien:
          type: string